sabiC

+++ 0x01-variables_if_else_while/9-print_comb.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	int d = 48;
 	int dcomaspace [] = {44, 32};
-	int i;
+	unsigned long int i;
 
 	while (d <= 56)
 	{
